Prefinal

+++ src/Pages/MyCart.jsx
@@ -209,7 +209,7 @@ const MyCart = () => {
 
     const deleteHandler = (product) => {
         // alert(productId)
-        dispatch(deleteItemFromCart(product))
+
         Swal.fire({
             title: 'Estas seguro?',
             text: "Vas a eliminar un producto de tu Carrito!",
@@ -219,6 +219,16 @@ const MyCart = () => {
             cancelButtonColor: '#E90000',
             confirmButtonText: 'Eliminar',
             cancelButtonText: 'Cancelar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                dispatch(deleteItemFromCart(product))
+
+                Swal.fire(
+                    'Eliminado!',
+                    'Tu categoría fue eliminada.',
+                    'success'
+                )
+            }
         })
     }
 
@@ -228,9 +238,31 @@ const MyCart = () => {
 
     const decrementHandler = (product) => {
         if (product.quantity === 1) {
-            dispatch(deleteItemFromCart(product))
+            Swal.fire({
+                title: 'Estas seguro?',
+                text: "Vas a eliminar un producto de tu Carrito!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#7FFF00',
+                cancelButtonColor: '#E90000',
+                confirmButtonText: 'Eliminar',
+                cancelButtonText: 'Cancelar'
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    dispatch(deleteItemFromCart(product))
+
+                    Swal.fire(
+                        'Eliminado!',
+                        'Tu categoría fue eliminada.',
+                        'success'
+                    )
+                }
+            })
+
         }
-        dispatch(decrementQuantity(product))
+        if (product.quantity > 1) {
+            dispatch(decrementQuantity(product))
+        }
     }
